chore: restructure folder layout for better organization
