--- todo-list/src/Divider/Divider.tsx
@@ -1,9 +1,0 @@
-import styles from './Divider.module.css';
-
-const Divider = () => {
-    return (
-        <hr className={styles.container}/>
-    )
-};
-
-export default Divider;
