+++ todo-list/src/app/index.tsx
@@ -1,0 +1,12 @@
+import "../App.css";
+import Todo from "../pages/Todo";
+
+function App() {
+  return (
+    <main className="App">
+      <Todo />
+    </main>
+  );
+}
+
+export default App;
