+++ todo-list/src/entities/ui/index.ts
@@ -1,0 +1,1 @@
+export * from "../../features/todo/ui/TodoItem";
