+++ todo-list/src/shared/types/index.ts
@@ -1,0 +1,5 @@
+export type TodoType = {
+    id: number
+    text: string
+    isChecked: boolean
+  }
