+++ todo-list/src/widgets/Divider/index.tsx
@@ -1,0 +1,7 @@
+import styles from './styles/Divider.module.css';
+
+export const Divider = () => {
+    return (
+        <hr className={styles.container}/>
+    )
+};
